+++ src/main/java/ru/izerus/sibgeology/geology/utils/RandomUtils.java
@@ -6,17 +6,17 @@ public class RandomUtils {
 
 	private static final Random rand = new Random(System.currentTimeMillis());
 
-	public static float getInaccurateValue(float value, float inaccuracity) {
+	public static double getInaccurateValue(double value, double inaccuracity) {
 		if (inaccuracity <= 0)
 			return value;
-		float min = value - (value * inaccuracity);
+		double min = value - (value * inaccuracity);
 		min = min > 0 ? min : 0;
-		float max = value + (value * inaccuracity);
-		return min + rand.nextFloat() * (max - min);
+		double max = value + (value * inaccuracity);
+		return min + rand.nextDouble() * (max - min);
 	}
 
-	public static boolean isChance(float chance) {
-		return rand.nextFloat() <= chance;
+	public static boolean isChance(double chance) {
+		return rand.nextDouble() <= chance;
 	}
 
 }
